Write time in microseconds and don't miss last one
We want to use csv files as an indicators of abs time range for
recording because ttyrec contains only relative durations.

We need to use nested goroutine to watch for cancel call and
properly handle last write + close.

This way we keep file descriptors private and the function is
responsible for closing its dependencies.