Increase metrics granularity to determine sessions switch
If used do a session switch the worst scenario is when he did actions
in both sessions within the same metrics interval. In that case the
most active will win which isn't always the last one.

To reduce that probability and d not spend tons of time on refactoring
we can increase resolution of a metrics to handle that case.