Handle args and terminal state properly
We should restore terminal state after switching to RAW mode, otherwise parent shell will be broken. That may lead to UI artifacts.

Additionally, we should respect passed args and forward it to the child shell.

If `QUDOSH_SHELL` is unset we can use binary name, this way we can store qudosh as `zsh` in custom folder. To handle same name we need to call child shell as abs path.